commit lesson 3 configure

+++ Application/Services/PersonService.cs
@@ -2,41 +2,77 @@ using System;
 using System.Collections.Generic;
 using Application.Interfaces;
 using Domain.Entities;
+using Application.DTOs;
+using Application.MappingProfiles;
+using AutoMapper;
 
 namespace Application.Services
 {
-    public class PersonService: IPersonRepository
+    public class PersonService<PersonCreateResponse>
     {
         public readonly IPersonRepository _personRepository;
+        public readonly IMapper _mapper;
         // TODO: реализовать crud
         
-        public PersonService(IPersonRepository personRepository)
+        public PersonService(IPersonRepository personRepository, IMapper mapper)
         {
             _personRepository = personRepository;
+            _mapper = mapper;
         }
 
-        public Person GetById(Guid id)
+        public PersonGetByResponse GetById(Guid id)
         {
-            return _personRepository.GetById(id);
+            var person = _personRepository.GetById(id);
+            var model = _mapper.Map<PersonGetByResponse>(person);
+
+            return model;
         }
 
-        public List<Person> Get()
+        public List<PersonGetAllResponse> Get()
         {
-            return _personRepository.Get();
+            var persons = _personRepository.GetAll();
+            var model = _mapper.Map<List<PersonGetAllResponse>>(persons);
+            return model;
         }
-
-        public Person Create(Person person)
+//TODO resp req для каждого dto
+        public PersonCreateResponse Create(PersonCreateRequest personDto)
         {
-            return _personRepository.Create(person);
+            var person = _mapper.Map<Person>(personDto); // Преобразуем DTO в сущность Person
+            var createdPerson = _personRepository.Create(person); // Создаем пользователя, используя репозиторий
+            var createdDto = _mapper.Map<PersonCreateResponse>(createdPerson); // Преобразуем созданную сущность Person в DTO
+            return createdDto; // Возвращаем DTO
         }
-
-        public Person Update(Person person)
+        
+        public PersonUpdateResponse Update(PersonUpdateRequest personUpdateRequest)
         {
-            return _personRepository.Update(person);
+            var person = _personRepository.GetById(personUpdateRequest.Id);
+
+            person.Update(personUpdateRequest.FirstName, personUpdateRequest.LastName, personUpdateRequest.MiddleName,
+                personUpdateRequest.PhoneNumber);
+            var personResponse = _personRepository.Update(person);
+            var response = _mapper.Map<PersonUpdateResponse>(personResponse);
+            return response;
+            /*var existingPerson = _personRepository.GetById(id);
+            if (existingPerson == null)
+            {
+                throw new Exception("Пользователь не найден");
+            }
+
+            _mapper.Map(personDto, existingPerson);
+            _personRepository.Update(existingPerson);
+
+            var updatedDto = _mapper.Map<PersonGetByResponse>(existingPerson);
+            return updatedDto;*/
         }
 
         public bool Delete(Guid id)
         {
+            var existingPerson = _personRepository.GetById(id);
+            if (existingPerson == null)
+            {
+                return false;
+            }
+
             return _personRepository.Delete(id);
         }
 
@@ -44,5 +80,6 @@ namespace Application.Services
         {
             return _personRepository.GetCustomFields();
         }
+
     }
 }
